fix gameroom router add authorization and send userId with login router
