+++ user/model.js
@@ -2,7 +2,7 @@ const Sequelize = require("sequelize");
 const db = require("../db.js");
 
 const User = db.define("user", {
-  username: { type: Sequelize.STRING, allowNull: false },
+  username: { type: Sequelize.STRING, allowNull: true },
   email: { type: Sequelize.STRING, unique: true, allowNull: false },
   password: { type: Sequelize.STRING, allowNull: false }
 });
